Move SandboxSpecExtensions to it's own file and will add WaitSync method
I'm adding a SyncCompleted Event Handler to Presence to help with the
WaitSync implementation
