+++ src/IO.Ably.Tests.Shared/IO.Ably.Tests.Shared.projitems
@@ -105,6 +105,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Rest\RestSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rest\SandboxSpec.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rest\RequestSandBoxSpecs.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Rest\SandboxSpecExtension.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rest\StatsSandBoxSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Rest\StatsSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Infrastructure\TestLogger.cs" />
