+++ src/IO.Ably.Tests.Shared/IO.Ably.Tests.Shared.shproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup Label="Globals">
     <ProjectGuid>1609f12c-8216-4e7c-ade0-240bfe160242</ProjectGuid>
